Move gen_config to own file
