+++ src/pbfalcon/tasks/basic.py
@@ -34,49 +34,6 @@ def FT(file_type, basename):
 RDJ = FT(FC_BASH, 'run_daligner_jobs.sh')
 
 
-# temporary defaults for lambda
-# see: http://bugzilla.nanofluidics.com/show_bug.cgi?id=28896
-_defaults_for_task_falcon_get_config = """\
-falcon_sense_option = --output_multi --min_idt 0.70 --min_cov 4 --local_match_count_threshold 2 --max_n_read 200 --n_core 6
-length_cutoff = 12000
-length_cutoff_pr = 12000
-pa_DBsplit_option = -x500 -s50
-pa_HPCdaligner_option = -v -dal4 -t16 -e.70 -l1000 -s1000
-pa_concurrent_jobs = 32
-overlap_filtering_setting = --max_diff 100 --max_cov 50 --min_cov 1 --bestn 10 --n_core 24
-ovlp_HPCdaligner_option = -v -dal4 -t32 -h60 -e.96 -l500 -s1000
-ovlp_DBsplit_option = -x500 -s50
-ovlp_concurrent_jobs = 32
-"""
-# also see:
-#   https://dazzlerblog.wordpress.com/command-guides/daligner-command-reference-guide/
-#   https://dazzlerblog.wordpress.com/2014/06/01/the-dazzler-db/
-#   https://github.com/PacificBiosciences/FALCON/wiki/Manual
-
-def sorted_str(s):
-    return '\n'.join(sorted(s.splitlines()))
-
-_defaults_for_task_falcon_get_config = sorted_str(_defaults_for_task_falcon_get_config)
-
-def _get_defaults_for_task_falcon_get_config():
-    #result = pbfalcon.ini2dict(_defaults_for_task_falcon_get_config)
-    result = dict()
-    result.update({
-        'GenomeLength_int': '5000000',
-        'CoresMax_int': '40',
-        'FalconAdvanced_str': pbfalcon.ini2option_text(_defaults_for_task_falcon_get_config),
-    })
-    return result
-
-@registry('task_falcon_get_config', '0.0.0', [FC_FOFN], [FC_CONFIG],
-        options=_get_defaults_for_task_falcon_get_config(),
-        is_distributed=False)
-def run_rtc(rtc):
-  with cd(os.path.dirname(rtc.task.output_files[0])):
-    return pbfalcon.run_falcon_get_config(rtc.task.input_files,
-                                          rtc.task.output_files,
-                                          rtc.task.options)
-
 @registry('task_falcon_config_get_fasta', '0.0.0', [FC_CONFIG], [FC_FOFN], is_distributed=False)
 def run_rtc(rtc):
   with cd(os.path.dirname(rtc.task.output_files[0])):
