+++ src/pbfalcon/tusks.py
@@ -50,92 +50,6 @@ def _get_config_from_json_fileobj(ifs_json):
     say('JSON=\n%s' %i_json[:1024]) # truncated
     return json.loads(i_json)
 
-def _options_dict_with_base_keys(options_dict):
-    """Remove leading namespaces from key names,
-    in a copy of options_dict.
-    """
-    new_dict = dict()
-    for key, val in options_dict.items():
-        new_dict[key.split('.')[-1]] = val
-    return new_dict
-
-def _gen_config(options_dict):
-    """Generate ConfigParser object from dict.
-    """
-    cfg = support.parse_config('')
-    sec = "General"
-    cfg.add_section(sec)
-    for key, val in options_dict.items():
-        # Strip leading and trailing ws, b/c the pbsmrtpipe
-        # misinterprets XML as a data-interchanges language.
-        # (It is only mark-up, so ws is never meaningful.)
-        # Also, we want only strings; hopefully, we can fix
-        # the TC later to drop the type-info (e.g. integer).
-        cfg.set(sec, key, str(val).strip())
-    return cfg
-
-def _write_config(config, config_fn):
-    with open(config_fn, 'w') as ofh:
-        config.write(ofh)
-
-def ini2dict(ini_text):
-    ifp = StringIO.StringIO('[General]\n' + ini_text)
-    cp = configparser.ConfigParser()
-    cp.readfp(ifp)
-    return dict(cp.items('General'))
-
-re_semicolon = re.compile(r'\s*;+\s*')
-
-def option_text2ini(option_text):
-    # Basically, just translate semicolons into linefeeds.
-    return re_semicolon.sub('\n', option_text)
-
-re_newline = re.compile(r'\s*\n\s*', re.MULTILINE)
-
-def ini2option_text(ini):
-    # Basically, just translate linefeeds into semicolons.
-    return re_newline.sub(';', ini)
-
-def get_falcon_overrides(cfg_content, OPTION_CFG=OPTION_CFG):
-    """options keys are bare (no namespaces)
-    """
-    if '\n' in cfg_content:
-        log.error('linefeed found in option "%s", which is ok here but should have been prevented earler' %(
-            OPTION_CFG))
-        cfg_content = ini2option_text(cfg_content)
-    if cfg_content.strip().startswith('['):
-        log.error('Option "%s" seems to have .ini-style [brackets], which is an error. It is not really a .ini file.' %(
-            OPTION_CFG))
-        # Try to strip the first line, and hope there are no others.
-        cfg_content = cfg_content[cfg_content.index(']'):]
-    ini = option_text2ini(cfg_content)
-    log.info(ini)
-    # Now, parse the overrides, but skip it on any error.
-    try:
-        overrides = ini2dict(ini)
-    except Exception as exc:
-        log.exception('For option "%s" (for overrides) we had a problem parsing its contents:\n%s' %(
-            OPTION_CFG, cfg_content))
-        overrides = dict()
-    return overrides
-
-def run_falcon_gen_config(input_files, output_files, options):
-    """Generate a config-file from options.
-
-    TODO(CD):
-    Eventually, use GenomeSize and ParallelTasksMax too.
-    Also, validate cfg, in case of missing options.
-    """
-    i_fofn_fn, = input_files
-    o_cfg_fn, = output_files
-    options = _options_dict_with_base_keys(options)
-    if OPTION_CFG in options:
-        overrides = get_falcon_overrides(options[OPTION_CFG], OPTION_CFG)
-        options.update(overrides)
-    config = _gen_config(options)
-    with cd(os.path.dirname(i_fofn_fn)):
-        return _write_config(config, o_cfg_fn)
-
 def run_falcon_config_get_fasta(input_files, output_files):
         i_config_fn, = input_files
         o_fofn_fn, = output_files
